+++ routes/test.ts
@@ -1,4 +1,3 @@
-import TestController from "@/app/controllers/TestController";
 import TestMiddleware from "@/app/middlewares/TestMiddleware";
 import LoggerMiddleware from "@/app/middlewares/LoggerMiddleware";
 import Router from "@/utils/Router";
@@ -8,23 +7,11 @@ export default Router.prefix("test")
         new TestMiddleware(),
         new LoggerMiddleware()
     )
-    .group({
-        "get": {
-            GET: new TestController().get
-        },
-        "detail/:id": {
-            GET: new TestController().detail
-        },
-        "add": {
-            POST: new TestController().add
-        },
-        "edit": {
-            POST: new TestController().edit
-        },
-        "delete/:id": {
-            DELETE: new TestController().delete
-        },
-        "restore/:id": {
-            GET: new TestController().restore
-        }
-    });
+    .group([
+        Router.get("get", "TestController@get"),
+        Router.get("detail/:id", "TestController@detail"),
+        Router.post("add", "TestController@add"),
+        Router.post("edit", "TestController@edit"),
+        Router.delete("delete/:id", "TestController@delete"),
+        Router.get("restore/:id", "TestController@restore")
+    ]);