+++ types/middleware.d.ts
@@ -1,7 +1,4 @@
-import {BunRequest} from "bun";
-
 declare global {
-    type HandlerType = (request: BunRequest) => Promise<Response>;
     type MiddlewareType = (handler: HandlerType) => HandlerType;
 
     interface Middleware {