+++ CHANGELOG.md
@@ -3,7 +3,29 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [v0.1.2](https://github.com/crenata/bejibun/compare/v0.1.0...v0.1.2) - 2025-10-10
+## [v0.1.3](https://github.com/crenata/bejibun/compare/v0.1.2...v0.1.3) - 2025-10-11
+
+### 🩹 Fixes
+
+### 📖 Changes
+New Features :
+- Adding router for http methods, e.g. `Router.get("get", "TestController@get")`
+- Now, router group support object and array
+
+Chore :
+- Rename `CorsMethodEnum` to `HttpMethodEnum`
+- Move some types to global
+- Adding `RouterInvalidException` exception
+
+### ❤️Contributors
+- Havea Crenata ([@crenata](https://github.com/crenata))
+- Ghulje ([@ghulje](https://github.com/ghulje))
+
+**Full Changelog**: https://github.com/crenata/bejibun/blob/master/CHANGELOG.md
+
+---
+
+## [v0.1.2](https://github.com/crenata/bejibun/compare/v0.1.1...v0.1.2) - 2025-10-10
 
 ### 🩹 Fixes